Extract the TestSparkActionWindow class into a separate file
Extract the TestSparkActionWindow class into a separate file for better modularity and maintainability. This separation improves code readability and simplifies the structure of TestSparkAction by isolating UI logic. It will also reduce potential git conflicts.
